chore(vm): Refactor CompileContext to its own file, create FinaliserStack types
